Created two separate components for log in and sign up
