+++ frontend/src/components/Footer.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import styled from 'styled-components';
+import styled from 'styled-components/macro';
 
 export const Footer = () => {
   return (
